python-backend/requirements.txt
fastapi[standard]
pydantic
python-dotenv
google-generativeai
ipython